pyproject.toml
[project]
name = "prompt-engineering"
version = "0.1.0"
description = "Prompt engineering practice 01"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
]
